got sqee compiling with MSVC
some huge improvements to cmake stuff
rpg doesn't build because of bad c++17 support

+++ include/sqee/dop/Classes.hpp
@@ -196,7 +196,7 @@ template <class Type> inline
 uint Table<Type>::find(int32_t id) const
 {
     const auto iter = algo::find(mIds, id);
-    return std::distance(mIds.begin(), iter);
+    return uint(std::distance(mIds.begin(), iter));
 }
 
 //============================================================================//
